requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.2.1
requests==2.32.3
selenium==4.21.0
seletools==1.5.0
Faker~=24.3.0